+++ NEWS.md
@@ -1,6 +1,9 @@
 
 # vctrs (development version)
 
+* vctrs now supports the `data.table` class. The common type of a data
+  frame and a data table is a data table.
+
 * `new_vctr()` now always appends a base `"list"` class to list `.data` to
   be compatible with changes to `vec_is_list()`. This affects `new_list_of()`,
   which now returns an object with a base class of `"list"`.